Composition relationship - Move Schedule to CourseService
