--- HAN.Services/Interfaces/IScheduleService.cs
@@ -1,10 +1,0 @@
-﻿using HAN.Services.DTOs;
-
-namespace HAN.Services.Interfaces;
-
-public interface IScheduleService
-{
-        public ScheduleDto AddSchedule(ScheduleDto schedule, int courseId);
-        public ScheduleDto GetScheduleById(int id);
-        public ScheduleDto UpdateSchedule(int courseId, ScheduleDto scheduleDto);
-}
